refactor(reducer): move logic away from index.js
